implement redis listener thread and rework auth token storage
We used to just store the access token, but that doesn't work because we
need the username. Furthermore, the oauth web flow doesn't even give us
the username! In order to get it we need to make an API call that is not
supported in any version of the library lower than 0.14, which we depend
on as of a previous commit.

Now, upon completion of the oauth flow, we make the API call to query
the authenticated user's data and store a GithubConf, (serialized to
JSON,) which contains the username, the bearer token, and the gitmail
repo name, to Redis.

In a future revision, we should store this information in a database.